機能: アノテーションデータのインポート元を変更し、型定義を更新
- `formatTweetQueryParams.ts`と`renderAnnotation.ts`で`MarkerAnnotationData`のインポート元を`common/GetAnnotationsType`に変更しました。
- `GetMissionType.ts`と`GetNewsType.ts`で`GetFireData`と`GetNewsData`の型を`GetAnnotationsType`に更新しました。
