+++ src/utils/function/map/renderAnnotation.ts
@@ -1,11 +1,7 @@
-import {
-  GetNewsData,
-  MapInstance,
-  MapkitInstance,
-  MarkerAnnotationData,
-} from "@/utils/type/api/GetNewsType"
+import { GetNewsData, MapInstance, MapkitInstance } from "@/utils/type/api/GetNewsType"
 import { RefObject } from "react"
 import { categoryStyleMap } from "./categoryStyleMap"
+import { MarkerAnnotationData } from "@/utils/type/api/common/GetAnnotationsType"
 
 /**
  * マーカー（アノテーション）の表示をする関数
