+++ src/utils/type/api/GetMissionType.ts
@@ -1,5 +1,5 @@
-import { MapAnnotationType } from "../map/MapAnnotationType"
+import { GetAnnotationsType } from "./common/GetAnnotationsType"
 
 export const FIRE_API_ENDPOINT = `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/fire-trucks/missions`
 
-export type GetFireData = MapAnnotationType
+export type GetFireData = GetAnnotationsType
